created reusable logos editor page component and added affiliates editor page
